Serialise log arguments, and store and render as 'values'
